extract fuction validator to separate module; add tests for validate state
